Got Markers showing on MapFragment!
I got up to 20 pins showing up on the map. Need database feedback to
filter out some of them as google sees hamburgers as a type of sandwich
and added some sandwich places to the pins. I'll add the mobile blue
dot, indicating self position, in the afternoon. Also I still cannot get
this to run on my actual phone, but it works just fine on a emulator.